+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/InProcessResolver.java
@@ -7,6 +7,8 @@ import dev.openfeature.contrib.providers.flagd.resolver.process.model.FeatureFla
 import dev.openfeature.contrib.providers.flagd.resolver.process.storage.FlagStore;
 import dev.openfeature.contrib.providers.flagd.resolver.process.storage.Storage;
 import dev.openfeature.contrib.providers.flagd.resolver.process.storage.StorageState;
+import dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.Connector;
+import dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.file.FileConnector;
 import dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.grpc.GrpcStreamConnector;
 import dev.openfeature.contrib.providers.flagd.resolver.process.targeting.Operator;
 import dev.openfeature.contrib.providers.flagd.resolver.process.targeting.TargetingRuleException;
@@ -26,7 +28,8 @@ import java.util.function.Consumer;
 import static dev.openfeature.contrib.providers.flagd.resolver.process.model.FeatureFlag.EMPTY_TARGETING_STRING;
 
 /**
- * flagd in-process resolver. Resolves feature flags in-process. Flags are retrieved from {@link Storage}, where the
+ * flagd in-process resolver. Resolves feature flags in-process. Flags are
+ * retrieved from {@link Storage}, where the
  * {@link Storage} maintain flag configurations obtained from known source.
  */
 @Slf4j
@@ -41,8 +44,11 @@ public class InProcessResolver implements Resolver {
      * Initialize an in-process resolver.
      */
     public InProcessResolver(FlagdOptions options, Consumer<ProviderState> stateConsumer) {
-        // currently we support gRPC connector
-        this.flagStore = new FlagStore(new GrpcStreamConnector(options));
+        final Connector connector = options.isOffline()
+                ? new FileConnector(options.getOfflineFlagSourcePath())
+                : new GrpcStreamConnector(options);
+
+        this.flagStore = new FlagStore(connector);
         this.deadline = options.getDeadline();
         this.stateConsumer = stateConsumer;
         this.operator = new Operator();
@@ -97,7 +103,7 @@ public class InProcessResolver implements Resolver {
      * Resolve a boolean flag.
      */
     public ProviderEvaluation<Boolean> booleanEvaluation(String key, Boolean defaultValue,
-                                                         EvaluationContext ctx) {
+            EvaluationContext ctx) {
         return resolve(Boolean.class, key, ctx);
     }
 
@@ -105,7 +111,7 @@ public class InProcessResolver implements Resolver {
      * Resolve a string flag.
      */
     public ProviderEvaluation<String> stringEvaluation(String key, String defaultValue,
-                                                       EvaluationContext ctx) {
+            EvaluationContext ctx) {
         return resolve(String.class, key, ctx);
     }
 
@@ -113,7 +119,7 @@ public class InProcessResolver implements Resolver {
      * Resolve a double flag.
      */
     public ProviderEvaluation<Double> doubleEvaluation(String key, Double defaultValue,
-                                                       EvaluationContext ctx) {
+            EvaluationContext ctx) {
         return resolve(Double.class, key, ctx);
     }
 
@@ -121,7 +127,7 @@ public class InProcessResolver implements Resolver {
      * Resolve an integer flag.
      */
     public ProviderEvaluation<Integer> integerEvaluation(String key, Integer defaultValue,
-                                                         EvaluationContext ctx) {
+            EvaluationContext ctx) {
         return resolve(Integer.class, key, ctx);
     }
 
@@ -142,7 +148,7 @@ public class InProcessResolver implements Resolver {
     }
 
     private <T> ProviderEvaluation<T> resolve(Class<T> type, String key,
-                                              EvaluationContext ctx) {
+            EvaluationContext ctx) {
         final FeatureFlag flag = flagStore.getFlag(key);
 
         // missing flag
@@ -154,7 +160,6 @@ public class InProcessResolver implements Resolver {
         if ("DISABLED".equals(flag.getState())) {
             throw new FlagNotFoundError("flag: " + key + " is disabled");
         }
-
 
         final Object resolvedVariant;
         final String reason;
@@ -186,7 +191,13 @@ public class InProcessResolver implements Resolver {
             log.debug(message);
             throw new TypeMismatchError(message);
         }
-
+        if (value instanceof Integer && type == Double.class) {
+            // if this is an integer and we are trying to resolve a double, convert
+            value = ((Integer) value).doubleValue();
+        } else if (value instanceof Double && type == Integer.class) {
+            // if this is a double and we are trying to resolve an integer, convert
+            value = ((Double) value).intValue();
+        }
         if (!type.isAssignableFrom(value.getClass()) || !(resolvedVariant instanceof String)) {
             String message = "returning default variant for flagKey: %s, type not valid";
             log.debug(String.format(message, key));